+++ src/plugins/lokijs/rx-storage-instance-loki.ts
@@ -80,6 +80,7 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
     public readonly instanceId = instanceId++;
 
     public readonly leaderElector?: LeaderElector;
+    private closed = false;
 
     constructor(
         public readonly databaseName: string,
@@ -141,7 +142,11 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
      * If the local state must be used, that one is returned.
      * Returns false if a remote instance must be used.
      */
-    private async mustUseLocalState(): Promise<LokiLocalDatabaseState | false> {
+    public async mustUseLocalState(): Promise<LokiLocalDatabaseState | false> {
+        if (this.closed) {
+            return false;
+        }
+
         if (this.internals.localState) {
             return this.internals.localState;
         }
@@ -164,6 +169,7 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
             leaderElector.isLeader &&
             !this.internals.localState
         ) {
+
             // own is leader, use local instance
             this.internals.localState = createLokiLocalState({
                 databaseName: this.databaseName,
@@ -711,6 +717,7 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
         return this.changes$.asObservable();
     }
     async close(): Promise<void> {
+        this.closed = true;
         this.changes$.complete();
         OPEN_LOKIJS_STORAGE_INSTANCES.delete(this);
 
@@ -738,6 +745,7 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
         }
         localState.databaseState.database.removeCollection(this.collectionName);
         localState.databaseState.database.removeCollection(localState.changesCollection.name);
+        this.closed = true;
     }
 }
 
@@ -825,6 +833,7 @@ export async function createLokiStorageInstance<RxDocType>(
         await internals.localState;
     }
 
+
     const instance = new RxStorageInstanceLoki(
         params.databaseName,
         params.collectionName,
@@ -835,6 +844,14 @@ export async function createLokiStorageInstance<RxDocType>(
         params.idleQueue,
         params.broadcastChannel
     );
+
+    /**
+     * Directly create the localState if the db becomes leader.
+     */
+    if (params.broadcastChannel) {
+        const leaderElector = getLeaderElectorByBroadcastChannel(params.broadcastChannel);
+        leaderElector.awaitLeadership().then(() => instance.mustUseLocalState());
+    }
 
     return instance;
 }